+++ src/win32/screen/drawdds.cpp
@@ -74,7 +74,7 @@ bool WinDrawDDS::Init(HWND hwindow, uint32_t w, uint32_t h, GUID* drv) {
   guimode = true;
   SetGUIMode(false);
   SendMessage(hwnd, WM_M88_CLIPCURSOR, CLIPCURSOR_WINDOW, 0);
-  status |= static_cast<uint32_t>(Draw::Status::kShouldRefresh);
+  status |= Draw::Status::kShouldRefresh;
 
   return true;
 }
@@ -259,7 +259,7 @@ bool WinDrawDDS::Lock(uint8_t** pimage, int* pbpl) {
 //  画面イメージの使用終了
 //
 bool WinDrawDDS::Unlock() {
-  status &= ~static_cast<uint32_t>(Draw::Status::kShouldRefresh);
+  status &= ~Draw::Status::kShouldRefresh;
   return true;
 }
 
@@ -333,7 +333,7 @@ bool WinDrawDDS::RestoreSurface() {
     return false;
   }
 
-  status |= static_cast<uint32_t>(Draw::Status::kShouldRefresh);
+  status |= Draw::Status::kShouldRefresh;
   FillBlankArea();
   return true;
 }
@@ -369,7 +369,7 @@ void WinDrawDDS::SetGUIMode(bool newguimode) {
     if (guimode) {
       ddraw->FlipToGDISurface();
       ddsscrn->SetClipper(ddcscrn);
-      status |= static_cast<uint32_t>(Draw::Status::kShouldRefresh);
+      status |= Draw::Status::kShouldRefresh;
     } else {
       ddsscrn->SetClipper(0);
 
@@ -381,7 +381,7 @@ void WinDrawDDS::SetGUIMode(bool newguimode) {
       rectsrc.right = width;
       rectsrc.bottom = height;
 
-      status |= static_cast<uint32_t>(Draw::Status::kShouldRefresh);
+      status |= Draw::Status::kShouldRefresh;
       Update(ddsscrn, rectsrc);
     }
   }
